refactor noun phrase to english
