+++ code/drasil-lang/lib/Language/Drasil/Chunk/DefinedQuantity.hs
@@ -23,7 +23,7 @@ import Language.Drasil.Chunk.UnitDefn (UnitDefn, unitWrapper,
   MayHaveUnit(getUnit))
 import Language.Drasil.Space (Space, HasSpace(..))
 import Language.Drasil.Stages (Stage (Implementation, Equational))
-import Language.Drasil.NounPhrase.Core (NP)
+import Language.Drasil.NaturalLanguage.English.NounPhrase.Core (NP)
 import Language.Drasil.Sentence (Sentence)
 
 -- | DefinedQuantityDict is the combination of a 'Concept' and a 'Quantity'.
