+++ code/drasil-lang/lib/Language/Drasil/Chunk/NamedIdea.hs
@@ -13,7 +13,7 @@ import Control.Lens ((^.), makeLenses, Lens')
 
 import Drasil.Database (mkUid, UID, HasUID(..), declareHasChunkRefs,
   Generically(..), IsChunk)
-import Language.Drasil.NounPhrase.Core (NP)
+import Language.Drasil.NaturalLanguage.English.NounPhrase.Core (NP)
 
 -- | A NamedIdea is a 'term' that we've identified (has a 'UID') as being worthy
 -- of naming.
