+++ code/drasil-lang/lib/Language/Drasil/Chunk/Relation.hs
@@ -14,7 +14,7 @@ import Language.Drasil.Chunk.Concept (ConceptChunk, dccWDS)
 import Language.Drasil.Classes (Express(..),
   ConceptDomain(..), Definition(..), Idea(..), NamedIdea(..))
 import Language.Drasil.ModelExpr.Lang (ModelExpr)
-import Language.Drasil.NounPhrase.Core (NP)
+import Language.Drasil.NaturalLanguage.English.NounPhrase.Core (NP)
 import Language.Drasil.Sentence (Sentence)
 
 -- | For a concept ('ConceptChunk') that also has a 'Relation' ('ModelExpr') attached.
