+++ code/drasil-lang/lib/Language/Drasil/Development.hs
@@ -14,7 +14,7 @@ module Language.Drasil.Development (
 ) where
 
 import Language.Drasil.Development.Sentence (toSent)
-import Language.Drasil.NounPhrase (NounPhrase(..))
-import Language.Drasil.NounPhrase.Core (NPStruct(..))
+import Language.Drasil.NaturalLanguage.English.NounPhrase (NounPhrase(..))
+import Language.Drasil.NaturalLanguage.English.NounPhrase.Core (NPStruct(..))
 import Language.Drasil.Sentence (sdep, shortdep, lnames, lnames')
 import Drasil.Database (showUID)
