+++ code/drasil-lang/lib/Language/Drasil/Development/Sentence.hs
@@ -24,8 +24,8 @@ import Language.Drasil.Chunk.NamedIdea (NamedIdea(term), Idea)
 import Language.Drasil.Sentence ((+:+), sParen, sentenceTerm,
   sentencePlural, sentenceShort)
 import qualified Language.Drasil.Sentence as S
-import qualified Language.Drasil.NounPhrase as NP
-import Language.Drasil.NounPhrase.Core (NPStruct(..))
+import qualified Language.Drasil.NaturalLanguage.English.NounPhrase as NP
+import Language.Drasil.NaturalLanguage.English.NounPhrase.Core (NPStruct(..))
 
 -- | Translate from NPStruct to Sentence
 toSent :: NPStruct -> S.Sentence
