+++ code/drasil-lang/lib/Language/Drasil/Sentence/Generators.hs
@@ -10,7 +10,7 @@ import Control.Lens ((^.))
 import Language.Drasil.Classes (NamedIdea(..), Quantity)
 import Language.Drasil.Development.Sentence (phrase)
 import Language.Drasil.Label.Type (Referable)
-import Language.Drasil.NounPhrase.Core (NP)
+import Language.Drasil.NaturalLanguage.English.NounPhrase.Core (NP)
 import Language.Drasil.Reference (refS)
 import Language.Drasil.Sentence (Sentence(S),(+:+), sParen, ch)
 import Language.Drasil.ShortName (HasShortName(..))
